+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0
+
+Smart Configuration Querying
+
+### Features:
+
+- Query Smart Configurations, set by the user in TrackZero Portal, using the groupId and an identifier
+
 ## 0.2.0
 
 Intital Release